added _memory_indexer() to base class `TraningDataset`
