+++ darts/utils/data/shifted_dataset.py
@@ -7,7 +7,8 @@ from typing import Union, Sequence, Optional, Tuple
 import numpy as np
 
 from ...timeseries import TimeSeries
-from .training_dataset import (PastCovariatesTrainingDataset,
+from .training_dataset import (TrainingDataset,
+                               PastCovariatesTrainingDataset,
                                FutureCovariatesTrainingDataset,
                                DualCovariatesTrainingDataset,
                                MixedCovariatesTrainingDataset,
@@ -381,7 +382,7 @@ class SplitCovariatesShiftedDataset(SplitCovariatesTrainingDataset):
         return past_target, past_covariate, future_covariate, future_target
 
 
-class GenericShiftedDataset:
+class GenericShiftedDataset(TrainingDataset):
     def __init__(self,
                  target_series: Union[TimeSeries, Sequence[TimeSeries]],
                  covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
@@ -469,13 +470,20 @@ class GenericShiftedDataset:
 
         # optionally, load covariates
         ts_covariate = self.covariates[ts_idx] if self.covariates is not None else None
+        cov_type = None
+        if self.covariates is not None:
+            cov_type = self.FUTURE_COV_TYPE if self.shift_covariates else self.PAST_COV_TYPE
 
         # get all indices for the current sample
         past_start, past_end, future_start, future_end, cov_start, cov_end = \
-            self._memory_indexer(ts_idx,
-                                 end_of_output_idx,
-                                 ts_target,
-                                 ts_covariate)
+            self._memory_indexer(ts_idx=ts_idx,
+                                 ts_target=ts_target,
+                                 shift=self.shift,
+                                 input_chunk_length=self.input_chunk_length,
+                                 output_chunk_length=self.output_chunk_length,
+                                 end_of_output_idx=end_of_output_idx,
+                                 ts_covariate=ts_covariate,
+                                 cov_type=cov_type)
 
         # extract sample target
         future_target = target_vals[future_start:future_end]
@@ -485,101 +493,15 @@ class GenericShiftedDataset:
         covariate = None
         if self.covariates is not None:
             raise_if_not(cov_end <= len(ts_covariate),
-                         f"The dataset contains {'future' if self.shift_covariates else 'past'} covariates "
+                         f"The dataset contains {cov_type} covariates "
                          f"that don't extend far enough into the future. ({idx}-th sample)")
 
             covariate = ts_covariate.values(copy=False)[cov_start:cov_end]
 
             raise_if_not(len(covariate) == (self.output_chunk_length if self.shift_covariates else
                                             self.input_chunk_length),
-                         f"The dataset contains {'future' if self.shift_covariates else 'past'} covariates "
+                         f"The dataset contains {cov_type} covariates "
                          f"whose time axis doesn't allow to obtain the input (or output) chunk relative to the "
                          f"target series.")
 
         return past_target, covariate, future_target
-
-    def _memory_indexer(self,
-                        ts_idx: int,
-                        end_of_output_idx: int,
-                        ts_target: TimeSeries,
-                        ts_covariate: TimeSeries) -> SampleIndexType:
-        """Returns the (start, end) indices for past target, future target and covariates (sub sets) of the current
-        sample `i` from `ts_idx`.
-
-        Works for all TimeSeries index types: pd.DatetimeIndex, pd.Int64Index, pd.RangeIndex
-
-        When `ts_idx` is observed for the first time, it stores the position of the sample `0` within the full target
-        time series and the (start, end) indices of all sub sets.
-        This allows to calculate the sub set indices for all future samples `i` by simply adjusting for the difference
-        between the positions of sample `i` and sample `0`.
-
-        Parameters
-        ----------
-        ts_idx
-            index of the current target TimeSeries.
-        ts_target
-            current target TimeSeries.
-        end_of_output_idx
-            the index where the output chunk of the current sample ends in `ts_target`.
-        ts_target
-            current target TimeSeries.
-        ts_covariate
-            current covariate TimeSeries.
-        """
-
-        cov_start, cov_end = None, None
-
-        # the first time ts_idx is observed
-        if ts_idx not in self._index_memory:
-            start_of_output_idx = end_of_output_idx - self.output_chunk_length
-            start_of_input_idx = start_of_output_idx - self.shift
-
-            # select forecast point and target period, using the previously computed indexes
-            future_start, future_end = start_of_output_idx, start_of_output_idx + self.output_chunk_length
-
-            # select input period; look at the `input_chunk_length` points after start of input
-            past_start, past_end = start_of_input_idx, start_of_input_idx + self.input_chunk_length
-
-            if self.covariates is not None:
-                start = future_start if self.shift_covariates else past_start
-                end = future_end if self.shift_covariates else past_end
-
-                # we need to be careful with getting ranges and indexes:
-                # to get entire range, full_range = ts[:len(ts)]; to get last index: last_idx = ts[len(ts) - 1]
-
-                # extract actual index value (respects datetime- and integer-based indexes; also from non-zero start)
-                start_time = ts_target.time_index[start]
-                end_time = ts_target.time_index[end - 1]
-
-                raise_if_not(start_time in ts_covariate.time_index and end_time in ts_covariate.time_index,
-                             f'Missing covariates; could not find {"future" if self.shift_covariates else "past"} '
-                             f'covariates in index value range: {start_time} - {end_time}.')
-
-                # extract the index position (index) from index value
-                cov_start = ts_covariate.time_index.get_loc(start_time)
-                cov_end = ts_covariate.time_index.get_loc(end_time) + 1
-
-            # store position of initial sample and all relevant sub set indices
-            self._index_memory[ts_idx] = {
-                'end_of_output_idx': end_of_output_idx,
-                'past_target': (past_start, past_end),
-                'future_target': (future_start, future_end),
-                'covariate': (cov_start, cov_end),
-            }
-        else:
-            # load position of initial sample and its sub set indices
-            end_of_output_idx_last = self._index_memory[ts_idx]['end_of_output_idx']
-            past_start, past_end = self._index_memory[ts_idx]['past_target']
-            future_start, future_end = self._index_memory[ts_idx]['future_target']
-            cov_start, cov_end = self._index_memory[ts_idx]['covariate']
-
-            # evaluate how much the new sample needs to be shifted, and shift all indexes
-            idx_shift = end_of_output_idx - end_of_output_idx_last
-            past_start += idx_shift
-            past_end += idx_shift
-            future_start += idx_shift
-            future_end += idx_shift
-            cov_start = cov_start + idx_shift if cov_start is not None else None
-            cov_end = cov_end + idx_shift if cov_end is not None else None
-
-        return past_start, past_end, future_start, future_end, cov_start, cov_end
