+++ darts/utils/data/training_dataset.py
@@ -8,9 +8,11 @@ from torch.utils.data import Dataset
 import numpy as np
 
 from typing import Tuple, Optional
-from ...logging import get_logger
+from ...logging import get_logger, raise_if_not
+from ...timeseries import TimeSeries
 
 logger = get_logger(__name__)
+SampleIndexType = Tuple[int, int, int, int, int, int]
 
 
 class TrainingDataset(ABC, Dataset):
@@ -56,6 +58,8 @@ class TrainingDataset(ABC, Dataset):
         and so we can get big performance gains when slicing by returning only numpy views of the data
         underlying the `TimeSeries`.
         """
+
+        self._index_memory = {}
         pass
 
     @abstractmethod
@@ -65,6 +69,102 @@ class TrainingDataset(ABC, Dataset):
     @abstractmethod
     def __getitem__(self, idx: int):
         pass
+
+    def _memory_indexer(self,
+                        ts_idx: int,
+                        ts_target: TimeSeries,
+                        shift: int,
+                        input_chunk_length: int,
+                        output_chunk_length: int,
+                        end_of_output_idx: int,
+                        ts_covariate: TimeSeries,
+                        cov_type: Optional[bool] = None) -> SampleIndexType:
+        """Returns the (start, end) indices for past target, future target and covariates (sub sets) of the current
+        sample `i` from `ts_idx`.
+
+        Works for all TimeSeries index types: pd.DatetimeIndex, pd.Int64Index, pd.RangeIndex
+
+        When `ts_idx` is observed for the first time, it stores the position of the sample `0` within the full target
+        time series and the (start, end) indices of all sub sets.
+        This allows to calculate the sub set indices for all future samples `i` by simply adjusting for the difference
+        between the positions of sample `i` and sample `0`.
+
+        Parameters
+        ----------
+        ts_idx
+            index of the current target TimeSeries.
+        ts_target
+            current target TimeSeries.
+        shift
+            The number of time steps by which to shift the output chunks relative to the input chunks.
+        input_chunk_length
+            The length of the emitted past series.
+        output_chunk_length
+            The length of the emitted future output series.
+        end_of_output_idx
+            the index where the output chunk of the current sample ends in `ts_target`.
+        ts_covariate
+            current covariate TimeSeries.
+        cov_type:
+            the type of covariate to extract: One of (None, 'past', 'future').
+        """
+
+        cov_start, cov_end = None, None
+
+        # the first time ts_idx is observed
+        if ts_idx not in self._index_memory:
+            start_of_output_idx = end_of_output_idx - output_chunk_length
+            start_of_input_idx = start_of_output_idx - shift
+
+            # select forecast point and target period, using the previously computed indexes
+            future_start, future_end = start_of_output_idx, start_of_output_idx + output_chunk_length
+
+            # select input period; look at the `input_chunk_length` points after start of input
+            past_start, past_end = start_of_input_idx, start_of_input_idx + input_chunk_length
+
+            if cov_type is not None:
+                start = future_start if cov_type == self.FUTURE_COV_TYPE else past_start
+                end = future_end if cov_type == self.FUTURE_COV_TYPE else past_end
+
+                # we need to be careful with getting ranges and indexes:
+                # to get entire range, full_range = ts[:len(ts)]; to get last index: last_idx = ts[len(ts) - 1]
+
+                # extract actual index value (respects datetime- and integer-based indexes; also from non-zero start)
+                start_time = ts_target.time_index[start]
+                end_time = ts_target.time_index[end - 1]
+
+                raise_if_not(start_time in ts_covariate.time_index and end_time in ts_covariate.time_index,
+                             f'Missing covariates; could not find {cov_type} covariates in index value range: '
+                             f'{start_time} - {end_time}.')
+
+                # extract the index position (index) from index value
+                cov_start = ts_covariate.time_index.get_loc(start_time)
+                cov_end = ts_covariate.time_index.get_loc(end_time) + 1
+
+            # store position of initial sample and all relevant sub set indices
+            self._index_memory[ts_idx] = {
+                'end_of_output_idx': end_of_output_idx,
+                'past_target': (past_start, past_end),
+                'future_target': (future_start, future_end),
+                'covariate': (cov_start, cov_end),
+            }
+        else:
+            # load position of initial sample and its sub set indices
+            end_of_output_idx_last = self._index_memory[ts_idx]['end_of_output_idx']
+            past_start, past_end = self._index_memory[ts_idx]['past_target']
+            future_start, future_end = self._index_memory[ts_idx]['future_target']
+            cov_start, cov_end = self._index_memory[ts_idx]['covariate']
+
+            # evaluate how much the new sample needs to be shifted, and shift all indexes
+            idx_shift = end_of_output_idx - end_of_output_idx_last
+            past_start += idx_shift
+            past_end += idx_shift
+            future_start += idx_shift
+            future_end += idx_shift
+            cov_start = cov_start + idx_shift if cov_start is not None else None
+            cov_end = cov_end + idx_shift if cov_end is not None else None
+
+        return past_start, past_end, future_start, future_end, cov_start, cov_end
 
 
 class PastCovariatesTrainingDataset(TrainingDataset, ABC):
